pyproject.toml
[project]
name = "django-itemlist"
version = "0.3.6"
description = "A customizable Django Admin ChangeList-like app for use outside of the admin."
authors = [
    {name = "Michel Fodje",email = "michel4j@gmail.com"}
]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/michel4j/django-itemlist"
requires-python = ">=3.10"
dependencies = [
    "django (>=4.2.20,<6.0.0)"
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry]
packages = [
    { include = "itemlist", from = "." },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-crisp-modals = ">=2025.3.10,<2026.0.0"